Fixes & Code is now runnable.
First, i fixed the input (the unembedding matrix was the wrong shape,
and the bias was none). In order to fix these, I transposed the
unembedding matrix, and in the case of a None type bias, a tensor of
all zeros was used instead.

Finally, the code was able to run locally on cpu. Next steps:
* Check in with Mansi
* Run code on Polaris GPUs